Include default input item logic in currencies hook
